headerfs: on Windows, close/truncate/reopen to avoid access denied
