+++ headerfs/file.go
@@ -46,39 +46,6 @@ func (h *headerStore) readRaw(seekDist int64) ([]byte, error) {
 	return rawHeader, nil
 }
 
-// singleTruncate truncates a single header from the end of the header file.
-// This can be used in the case of a re-org to remove the last header from the
-// end of the main chain.
-//
-// TODO(roasbeef): define this and the two methods above on a headerFile
-// struct?
-func (h *headerStore) singleTruncate() error {
-	// In order to truncate the file, we'll need to grab the absolute size
-	// of the file as it stands currently.
-	fileInfo, err := h.file.Stat()
-	if err != nil {
-		return err
-	}
-	fileSize := fileInfo.Size()
-
-	// Next, we'll determine the number of bytes we need to truncate from
-	// the end of the file.
-	var truncateLength int64
-	switch h.indexType {
-	case Block:
-		truncateLength = 80
-	case RegularFilter:
-		fallthrough
-	case ExtendedFilter:
-		truncateLength = 32
-	}
-
-	// Finally, we'll use both of these values to calculate the new size of
-	// the file and truncate it accordingly.
-	newSize := fileSize - truncateLength
-	return h.file.Truncate(newSize)
-}
-
 // readHeader reads a full block header from the flat-file. The header read is
 // determined by the hight value.
 func (h *BlockHeaderStore) readHeader(height int64) (*wire.BlockHeader, error) {
